backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0
python-dotenv==1.0.0

# Database
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
alembic==1.13.1

# Auth
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# Storage
boto3==1.34.20
Pillow==10.2.0

# AI/ML for body analysis
mediapipe==0.10.9
opencv-python==4.9.0.80
numpy==1.26.3

# Utilities
pandas==2.1.4
httpx==0.26.0
email-validator==2.1.0

# Dev
pytest==7.4.4
pytest-asyncio==0.23.3
black==23.12.1
ruff==0.1.11
